breaking into new files, make logs human-readable
