+++ mpi/mpi.go
@@ -3,7 +3,6 @@ package mpi
 import (
 	"bytes"
 	"crypto/md5"
-	"encoding/binary"
 	"encoding/json"
 	"fmt"
 	"net"
@@ -17,75 +16,7 @@ import (
 )
 
 func init() {
-	zap.ReplaceGlobals(zap.Must(zap.NewProduction()))
-}
-
-func SerializeWorld(world *MPIWorld) []byte {
-	// serialize the MPIWorld struct
-	// format: size, rank, IPPool, Port
-	// size: uint64
-	buf := make([]byte, 0)
-	sizebuf := make([]byte, 8)
-	binary.LittleEndian.PutUint64(sizebuf, world.size)
-	buf = append(buf, sizebuf...)
-
-	// rank: []uint64
-	for _, rank := range world.rank {
-		rankBuf := make([]byte, 8)
-		binary.LittleEndian.PutUint64(rankBuf, rank)
-		buf = append(buf, rankBuf...)
-	}
-
-	// IPPool: []string
-	for _, ip := range world.IPPool {
-		IPBuf := make([]byte, 0)
-		IPBuf = append(IPBuf, []byte(ip)...)
-		IPBuf = append(IPBuf, 0)
-		buf = append(buf, IPBuf...)
-	}
-
-	// Port: []uint64
-	for _, port := range world.Port {
-		portBuf := make([]byte, 8)
-		binary.LittleEndian.PutUint64(portBuf, port)
-		buf = append(buf, portBuf...)
-	}
-	return buf
-}
-
-func DeserializeWorld(buf []byte) *MPIWorld {
-	// deserialize the MPIWorld struct
-	// format: size, rank, IPPool, Port
-	// size: uint64
-	world := new(MPIWorld)
-	world.size = binary.LittleEndian.Uint64(buf[:8])
-	buf = buf[8:]
-
-	// rank: []uint64
-	world.rank = make([]uint64, world.size)
-	for i := uint64(0); i < world.size; i++ {
-		world.rank[i] = binary.LittleEndian.Uint64(buf[:8])
-		buf = buf[8:]
-	}
-
-	// IPPool: []string
-	world.IPPool = make([]string, world.size)
-	for i := uint64(0); i < world.size; i++ {
-		end := 0
-		for end < len(buf) && buf[end] != 0 {
-			end++
-		}
-		world.IPPool[i] = string(buf[:end])
-		buf = buf[end+1:]
-	}
-
-	// Port: []uint64
-	world.Port = make([]uint64, world.size)
-	for i := uint64(0); i < world.size; i++ {
-		world.Port[i] = binary.LittleEndian.Uint64(buf[:8])
-		buf = buf[8:]
-	}
-	return world
+	zap.ReplaceGlobals(zap.Must(zap.NewDevelopment()))
 }
 
 var (
@@ -264,17 +195,17 @@ func ReceiveBytes(size uint64, rank uint64) ([]byte, error) {
 }
 
 func GetHash(str string) {
-	fmt.Println(str + " Bytes sent: " + strconv.Itoa(int(BytesSent)))
-	fmt.Println(str + " Bytes received: " + strconv.Itoa(int(BytesReceived)))
-	fmt.Println(str + " Sent hash: " + fmt.Sprintf("%x", md5.Sum(sentBytes)))
-	fmt.Println(str + " Received hash: " + fmt.Sprintf("%x", md5.Sum(recvBytes)))
+	zap.L().Info(str + " Bytes sent: " + strconv.Itoa(int(BytesSent)))
+	zap.L().Info(str + " Bytes received: " + strconv.Itoa(int(BytesReceived)))
+	zap.L().Info(str + " Sent hash: " + fmt.Sprintf("%x", md5.Sum(sentBytes)))
+	zap.L().Info(str + " Received hash: " + fmt.Sprintf("%x", md5.Sum(recvBytes)))
 }
 
 func Close() {
-	fmt.Println("Bytes sent: " + strconv.Itoa(int(BytesSent)))
-	fmt.Println("Bytes received: " + strconv.Itoa(int(BytesReceived)))
-	fmt.Println("Sent hash: " + fmt.Sprintf("%x", md5.Sum(sentBytes)))
-	fmt.Println("Received hash: " + fmt.Sprintf("%x", md5.Sum(recvBytes)))
+	zap.L().Info("Bytes sent: " + strconv.Itoa(int(BytesSent)))
+	zap.L().Info("Bytes received: " + strconv.Itoa(int(BytesReceived)))
+	zap.L().Info("Sent hash: " + fmt.Sprintf("%x", md5.Sum(sentBytes)))
+	zap.L().Info("Received hash: " + fmt.Sprintf("%x", md5.Sum(recvBytes)))
 	if SelfRank == 0 {
 		time.Sleep(1 * time.Second)
 		for i := 1; i < len(DispatcherToWorkerTCPConn); i++ {
